#167: RegisterImageDetail ViewController - View 파일 분리 및 HeaderView/MainView 적용
